requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
docutils==0.15.2
idna==2.9
itsdangerous==1.1.0
jmespath==0.10.0
MarkupSafe==1.1.1
psycopg2==2.8.5
python-dateutil==2.8.1
python-dotenv==0.13.0
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
